Add initial code for basestation and node
The Code folder now contains the code for both Arduinos doing
communication through the powerline. The layout of the components
is detailed in the header of the files. This should enable
dynamic connection to the server via grabbing a UDP broadcast
of the basestation's IP address. The basestation will see the
client and send commands through the PLN device. The node device
sends a keep-alive 'Hello', which probably isn't necessary.

Also adds some files for Linksprite PLN testing.